% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRes.R
\name{parseRes}
\alias{parseRes}
\title{Parse results from MetaLearners()}
\usage{
parseRes(Meta.Res)
}
\arguments{
\item{Meta.Res}{Return object from \code{\link[=MetaLearners]{MetaLearners()}}}
}
\value{
A list of results where each element is a list of \code{ITE}, the estimated pairwise
treatment effect, and \code{Opt.trt}, the suggested optimal treatment.
}
\description{
Find the optimal treatment & pairwise treatment effect given results from MetaLearner()
}
\details{
Since results from \code{\link[=MetaLearners]{MetaLearners()}} only provides the marginal outcomes for many learners,
so it is necessary to translate the results into pairwise CATE/HTE/ITE and optimal treatment recommendations.
Note that for X-learner, we only return the estimated pairwise treatment effect but no optimal results.
People can choose their own method to figure out the optimal one based on pairwise results.
For simplex R-learner (Rsim) and AD-learning, only optimal treatment is provided as they are not designed to
estimate pairwise treatment effect.
}
