% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-package.R
\docType{package}
\name{stacking-package}
\alias{stacking-package}
\title{stacking: Experiments on Model Stacking Alternatives}
\description{
Supplies a dataset that arose from a set of experiments intended to benchmark the time-to-fit and predictive performance of a number of alternative model stacking methods (or "meta-learners") for use in the stacks package.
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
